ReplycA/requirements.txt
pyautogui
easyocr
edge-tts 
pynput 
playsound3
ollama 
requests 
beautifulsoup4 
duckduckgo_search 
opencv-python
pillow
SpeechRecognition
pyttsx3
PyAudio
mcp
transformers
datasets
torch
accelerate
python-dotenv
rich
redis
textual
tenacity
google-api-python-client
google-auth-oauthlib
google-auth-httplib2
openai-whisper
faster-whisper
librosa
soundfile
